diff + cast
